+++ pyccel/codegen/compiling/basic.py
@@ -95,7 +95,7 @@ class CompileObj:
         self._libs         = list(libs)
         self._libdir      = set(libdir)
         self._extra_compilation_tools = set(extra_compilation_tools)
-        self._dependencies = {a.module_target:a for a in dependencies}
+        self._dependencies = {getattr(a, 'module_target', a):a for a in dependencies}
         self._has_target_file = has_target_file
 
     def reset_folder(self, folder):