+++ tests/epyccel/test_loops.py
@@ -245,18 +245,15 @@ def test_for_expression(language):
 
 @pytest.mark.parametrize( 'language', (
         pytest.param("fortran", marks = [
-            pytest.mark.skip(reason="lists of lists not yet implemented in Fortran. See #2210"),
+            pytest.mark.skip(reason="lists of lists not yet implemented in Fortran. Types defined in other modules"),
             pytest.mark.fortran]
         ),
-        pytest.param("c", marks = [
-            pytest.mark.skip(reason="lists of lists not yet implemented in C. See #2210"),
-            pytest.mark.c]
-        ),
+        pytest.param("c", marks = pytest.mark.c),
         pytest.param("python", marks = pytest.mark.python)
     )
 )
 def test_for_lists_of_lists(language):
     f1 = loops.for_lists_of_lists
-    f2 = epyccel( f1, language = language )
+    f2 = epyccel( f1, language = language, flags='-Werror' )
 
     assert f1() == f2()